refactoring manual merge of pull request #851, insert copyright and rename files and class from RS_* to LC_*
