+++ librecad/src/src.pro
@@ -216,7 +216,7 @@ HEADERS += \
     actions/lc_actionfileexportmakercam.h \
     lib/engine/lc_rect.h \
     lib/printing/lc_printing.h \
-    actions/rs_actiondrawlinepolygon3.h
+    actions/lc_actiondrawlinepolygon3.h
 
 SOURCES += \
     lib/actions/rs_actioninterface.cpp \
@@ -305,7 +305,7 @@ SOURCES += \
     lib/engine/lc_rect.cpp \
     lib/engine/rs.cpp \
     lib/printing/lc_printing.cpp \
-    actions/rs_actiondrawlinepolygon3.cpp
+    actions/lc_actiondrawlinepolygon3.cpp
 
 # ################################################################################
 # Command
